+++ tests/Transformer/String/AsciiAlphanumericOnly/AsciiAlphanumericOnlyTransformerTest.php
@@ -2,28 +2,28 @@
 
 declare(strict_types=1);
 
-namespace ExtendsSoftware\ExaPHP\Transformer\String\AlphanumericOnly;
+namespace ExtendsSoftware\ExaPHP\Transformer\String\AsciiAlphanumericOnly;
 
 use PHPUnit\Framework\Attributes\CoversMethod;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
-#[CoversMethod(AlphanumericOnlyTransformer::class, 'transform')]
-class AlphanumericOnlyTransformerTest extends TestCase
+#[CoversMethod(AsciiAlphanumericOnlyTransformer::class, 'transform')]
+class AsciiAlphanumericOnlyTransformerTest extends TestCase
 {
     #[Test]
-    public function removesNonAlphanumericCharacters(): void
+    public function removesNonAsciiAlphanumericCharacters(): void
     {
-        $transformer = new AlphanumericOnlyTransformer();
+        $transformer = new AsciiAlphanumericOnlyTransformer();
         $value = $transformer->transform(' @ÁbC-12 3_Ⅱ'); // Container Unicode letter and number.
 
-        $this->assertSame('ÁbC123Ⅱ', $value);
+        $this->assertSame('bC123', $value);
     }
 
     #[Test]
     public function returnsValueWhenInputIsNotString(): void
     {
-        $transformer = new AlphanumericOnlyTransformer();
+        $transformer = new AsciiAlphanumericOnlyTransformer();
         $value = $transformer->transform(123);
 
         $this->assertSame(123, $value);